feat: separate question and answer routes into routers
